requirements.txt
python-dotenv==0.19.2
gunicorn==20.1.0

# Flask dependencies, plugins and extensions
Flask==2.0.3
